Create test for 'discover'.
