+++ package.json
@@ -3,6 +3,10 @@
   "version": "2.1.13",
   "description": "lightify node module",
   "license": "MIT",
+  "main": "lib/index.js",
+  "scripts": {
+    "test": "ava"
+  },
   "keywords": [
     "node",
     "osram",
@@ -19,8 +23,11 @@
     "node": ">=0.12.0"
   },
   "dependencies": {
-    "promise": ">=7.1.1",
-    "moment": ">=2.4.0"
+    "moment": ">=2.4.0",
+    "promise": ">=7.1.1"
   },
-  "readmeFilename": "README.md"
+  "readmeFilename": "README.md",
+  "devDependencies": {
+    "ava": "0.25.0"
+  }
 }
