+++ test/snapshots/index.js.md
@@ -1,0 +1,195 @@
+# Snapshot report for `test/index.js`
+
+The actual snapshot is saved in `index.js.snap`.
+
+Generated by [AVA](https://ava.li).
+
+## discover
+
+> Snapshot 1
+
+    {
+      request: '070000130100000001',
+      response: 'cb01011301000000000900e26b44520900002618840a0102049202010001648c0affffffff4131392054572036302030310000000000000000000000005b0f9ad40200002618840a0102049202010001648c0affffffff41313920545720363020303200000000000000000000000067f08f11e90000261884401000512102000001648c0a010000ff44696e696e6720526f6f6d0000000000480000000000000027eb24b7040000261884020102049202020001648c0a010000ff41313920545720363020303300000000000000000000000098c822ad040000261884020102049202020001648c0a010000ff413139205457203630203034000000000000000000000000aa537511e90000261884401000512102000001648c0a010000ff4b69746368656e000000000000000000020000000000000045ef6843050000261884020102049202040001648c0a010000ff41313920545720363020303500000000000000000000000025a38afe020000261884020102049202040001648c0a010000ff41313920545720363020303600000000000000000000000004d0bb12e90000261884401000512102000001648c0a010000ff4c6976696e6720526f6f6d00000000009100000000000000',
+      result: [
+        {
+          alpha: 255,
+          blue: 255,
+          brightness: 100,
+          firmware_version: 16909458,
+          friendlyMac: '4452090000261884',
+          green: 255,
+          groupid: 1,
+          id: 27618,
+          mac: -6.194884401344489e-289,
+          name: 'A19 TW 60 01',
+          online: 2,
+          red: 255,
+          status: 1,
+          temperature: 2700,
+          type: 10,
+        },
+        {
+          alpha: 255,
+          blue: 255,
+          brightness: 100,
+          firmware_version: 16909458,
+          friendlyMac: '9ad4020000261884',
+          green: 255,
+          groupid: 1,
+          id: 3931,
+          mac: -6.194884400956796e-289,
+          name: 'A19 TW 60 02',
+          online: 2,
+          red: 255,
+          status: 1,
+          temperature: 2700,
+          type: 10,
+        },
+        {
+          alpha: 255,
+          blue: 0,
+          brightness: 100,
+          firmware_version: 268456225,
+          friendlyMac: '8f11e90000261884',
+          green: 0,
+          groupid: 0,
+          id: 61543,
+          mac: -6.194884414708659e-289,
+          name: 'Dining Room',
+          online: 2,
+          red: 1,
+          status: 1,
+          temperature: 2700,
+          type: 64,
+        },
+        {
+          alpha: 255,
+          blue: 0,
+          brightness: 100,
+          firmware_version: 16909458,
+          friendlyMac: '24b7040000261884',
+          green: 0,
+          groupid: 2,
+          id: 60199,
+          mac: -6.19488440106938e-289,
+          name: 'A19 TW 60 03',
+          online: 2,
+          red: 1,
+          status: 1,
+          temperature: 2700,
+          type: 2,
+        },
+        {
+          alpha: 255,
+          blue: 0,
+          brightness: 100,
+          firmware_version: 16909458,
+          friendlyMac: '22ad040000261884',
+          green: 0,
+          groupid: 2,
+          id: 51352,
+          mac: -6.194884401067045e-289,
+          name: 'A19 TW 60 04',
+          online: 2,
+          red: 1,
+          status: 1,
+          temperature: 2700,
+          type: 2,
+        },
+        {
+          alpha: 255,
+          blue: 0,
+          brightness: 100,
+          firmware_version: 268456225,
+          friendlyMac: '7511e90000261884',
+          green: 0,
+          groupid: 0,
+          id: 21418,
+          mac: -6.194884414708635e-289,
+          name: 'Kitchen',
+          online: 2,
+          red: 1,
+          status: 1,
+          temperature: 2700,
+          type: 64,
+        },
+        {
+          alpha: 255,
+          blue: 0,
+          brightness: 100,
+          firmware_version: 16909458,
+          friendlyMac: '6843050000261884',
+          green: 0,
+          groupid: 4,
+          id: 61253,
+          mac: -6.194884401102106e-289,
+          name: 'A19 TW 60 05',
+          online: 2,
+          red: 1,
+          status: 1,
+          temperature: 2700,
+          type: 2,
+        },
+        {
+          alpha: 255,
+          blue: 0,
+          brightness: 100,
+          firmware_version: 16909458,
+          friendlyMac: '8afe020000261884',
+          green: 0,
+          groupid: 4,
+          id: 41765,
+          mac: -6.194884400966581e-289,
+          name: 'A19 TW 60 06',
+          online: 2,
+          red: 1,
+          status: 1,
+          temperature: 2700,
+          type: 2,
+        },
+        {
+          alpha: 255,
+          blue: 0,
+          brightness: 100,
+          firmware_version: 268456225,
+          friendlyMac: 'bb12e90000261884',
+          green: 0,
+          groupid: 0,
+          id: 53252,
+          mac: -6.194884414708932e-289,
+          name: 'Living Room',
+          online: 2,
+          red: 1,
+          status: 1,
+          temperature: 2700,
+          type: 64,
+        },
+      ],
+    
+
+## discoverZone
+
+> Snapshot 1
+
+    {
+      request: '0700001e0100000000',
+      response: '3f00011e01000000000300010044696e696e6720526f6f6d000000000002004b69746368656e00000000000000000003004c6976696e6720526f6f6d0000000000',
+      result: [
+        {
+          id: 1,
+          mac: 5e-324,
+          name: 'Dining Room',
+        },
+        {
+          id: 2,
+          mac: 1e-323,
+          name: 'Kitchen',
+        },
+        {
+          id: 3,
+          mac: 1.5e-323,
+          name: 'Living Room',
+        },
+      ],
+    }
